requirements.txt
python-dotenv
Flask
Flask-Mail